requirements.txt
Flask
Flask-SocketIO
Flask-Cors
python-dotenv
pandas
twilio
google-generativeai
eventlet
request